System.Net.WebHeaderCollection
